Test that at least one todo is returned when using filters
This commit just adds an extra assertion to the old tests. Everything still passes—I just wanted to make the tests a little stronger! 😊

Previously, we were asserting that all of the returned todos had the correct  owner (or status, or body, or category...) but we were not asserting that there were any todos present. If the database had given us an empty array, the tests would have passed vacuously.