Test that all the filters are case-insensitive
One test currently fails—status is currently case-sensitive. (At least, the 'status' query parameter in the URL is case-sensitive.)

It's possible that we don't strictly need status to be case-insensitive, but for the consistency and pleasant-to-use-ness of our API, I feel like making all the fields case-insensitive is the right call.

This commit addresses #1, #5, #6.